pyproject.toml
[project]
name = "excelagent"
version = "0.1.0"
description = "智能Excel操作代理 - 基于Agent Skills规范"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "openpyxl>=3.1",
    "pandas>=2.0",
    "Pillow>=10.0",
    "matplotlib>=3.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
include = ["agent*", "llm*", "tools*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
